+++ afat/tests/test_access.py
@@ -11,12 +11,10 @@ from django.urls import reverse
 # Alliance Auth
 from allianceauth.eveonline.models import EveCharacter
 
-# Alliance Auth (External Libs)
-from app_utils.testing import create_user_from_evecharacter
-
 # Alliance Auth AFAT
 from afat.tests import BaseTestCase
 from afat.tests.fixtures.load_allianceauth import load_allianceauth
+from afat.tests.fixtures.utils import create_user_from_evecharacter
 
 MODULE_PATH = "afat.views.statistics"