+++ afat/tests/test_views_fatlinks.py
@@ -19,13 +19,11 @@ from django.utils.datetime_safe import datetime
 # Alliance Auth
 from allianceauth.eveonline.models import EveCharacter
 
-# Alliance Auth (External Libs)
-from app_utils.testing import create_user_from_evecharacter
-
 # Alliance Auth AFAT
 from afat.models import Duration, Fat, FatLink, Log, get_hash_on_save
 from afat.tests import BaseTestCase
 from afat.tests.fixtures.load_allianceauth import load_allianceauth
+from afat.tests.fixtures.utils import create_user_from_evecharacter
 from afat.utils import get_main_character_from_user
 
 MODULE_PATH = "afat.views.fatlinks"